updated 4a variant to python 3

+++ mimic/scripts/ui_utils.py
@@ -4,17 +4,17 @@
 import maya.OpenMayaUI as omui
 
 # Use Qt.py to provide for back-compatibility from PySide2 to PySide
-from extern.Qt import QtWidgets
-from extern.Qt import QtGui
-from extern.Qt import QtCore
-from extern.Qt import QtCompat
+from Qt import QtWidgets
+from Qt import QtGui
+from Qt import QtCore
+from Qt import QtCompat
 
 
 def getMayaWindow():
     """
     """
     mayaMainWindowPtr = omui.MQtUtil.mainWindow() 
-    mayaMainWindow = QtCompat.wrapInstance(long(mayaMainWindowPtr), QtWidgets.QWidget) 
+    mayaMainWindow = QtCompat.wrapInstance(int(mayaMainWindowPtr), QtWidgets.QWidget)
 
     return mayaMainWindow
